base game functioning
