+++ game-of-life-app/src/components/Game.css
@@ -1,15 +1,15 @@
 body {
-    background-color: peachpuff;
+  background-color: peachpuff;
 }
-  
+
 .box {
-  display: inline-block;       
+  display: inline-block;
   border: 1px solid black;
   width: 15px;
   height: 15px;
   margin-left: -1px;
   margin-bottom: -1px;
-  }
+}
 
 .off {
   background-color: lightgray;
@@ -18,12 +18,17 @@ body {
 .on {
   background-color: green;
 }
-  
-.box:hover { 
-  background: #00CCFF;
+
+.box:hover {
+  background: #00ccff;
 }
 
 .center {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+.buttons {
   width: 600px;
   margin: auto;
   display: flex;
@@ -32,15 +37,10 @@ body {
   margin-top: 10px;
 }
 
-.grid {
-  width: 150px;
-  line-height: 0;
-  margin: auto;
-  box-shadow: 0px 0px 20px white;
-  margin-top: 20px;
-}
-  
-h1, h2, h3, h4 {
+h1,
+h2,
+h3,
+h4 {
   color: white;
   margin: auto;
   text-align: center;
